requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.12.3
boto3==1.34.100
botocore==1.34.162
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cssselect==1.3.0
cycler==0.12.1
datasets==2.19.1
dill==0.3.8
distlib==0.4.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.3.1
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.11
importlib_resources==6.5.2
jieba3k==0.35.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
kiwisolver==1.4.7
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.8.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.9.2
numpy==1.26.4
packaging==25.0
pandas==2.2.2
pillow==11.3.0
platformdirs==4.4.0
propcache==0.4.1
pyarrow==21.0.0
pyarrow-hotfix==0.7
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.31.0
requests-file==3.0.0
s3transfer==0.10.4
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8
sympy==1.14.0
tinysegmenter==0.3
tldextract==5.3.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
virtualenv==20.35.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
